algo

+++ libft/strtoint/strtoint.h
@@ -6,7 +6,7 @@
 /*   By: aghergut <aghergut@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:35:19 by aghergut          #+#    #+#             */
-/*   Updated: 2025/02/11 18:31:06 by aghergut         ###   ########.fr       */
+/*   Updated: 2025/02/12 18:04:58 by aghergut         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@
 //int to str
 char	*ft_itoa(int n);
 //str to int
-long	ft_atoi(const char *str);
+int		ft_atoi(const char *str);
 
 #endif
